+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - "Python: Lists" HackerRank challenge
+- "Python: Nested Lists" HackerRank challenge
 
 ## [0.1.0] - 2021-07-07
 ### Added